// src/enum/selectors.enum.ts
export enum SELECTORS {
    INPUT_COMBOBOX = 'input[role="combobox"]',
    OPTIONS_COMBOBOX_LINE_BUS_NUMBER = '.route-short-name',
    OPTIONS_COMBOBOX_ELEMENT = '.ng-dropdown-panel-items.scroll-host > div > div',
    ITINERARIES_CONTAINER = '.p-col-12.p-shadow-5.p-m-1.ng-star-inserted',
    LINE_BUS_START_POINT = 'span[ptooltip="Direção da viagem"]',
    DAILY_HOURS = '.trip-item.p-col-12.p-mt-1.ng-star-inserted',
    DAILY_WEEK = '.main-service-name.p-shadow-1.p-pl-1.p-pr-1.p-ml-1',
    DAILY_HOURS_INFO = '.departures.p-col-12 > a'
}